MAINT: avoid nested import by moving func to utils
