+++ mne_bids/write.py
@@ -36,7 +36,7 @@ from mne_bids.utils import (_write_json, _write_tsv, _read_events, _mkdir_p,
                             _check_key_val,
                             _parse_bids_filename, _handle_kind, _check_types,
                             _extract_landmarks, _parse_ext,
-                            _get_ch_type_mapping)
+                            _get_ch_type_mapping, make_bids_folders)
 from mne_bids.copyfiles import (copyfile_brainvision, copyfile_eeglab,
                                 copyfile_ctf, copyfile_bti, copyfile_kit)
 from mne_bids.read import reader
@@ -840,68 +840,6 @@ def make_bids_basename(subject=None, session=None, task=None,
     if isinstance(prefix, str):
         filename = op.join(prefix, filename)
     return filename
-
-
-def make_bids_folders(subject, session=None, kind=None, bids_root=None,
-                      make_dir=True, overwrite=False, verbose=False):
-    """Create a BIDS folder hierarchy.
-
-    This creates a hierarchy of folders *within* a BIDS dataset. You should
-    plan to create these folders *inside* the bids_root folder of the dataset.
-
-    Parameters
-    ----------
-    subject : str
-        The subject ID. Corresponds to "sub".
-    kind : str
-        The kind of folder being created at the end of the hierarchy. E.g.,
-        "anat", "func", etc.
-    session : str | None
-        The session for a item. Corresponds to "ses".
-    bids_root : str | None
-        The bids_root for the folders to be created. If None, folders will be
-        created in the current working directory.
-    make_dir : bool
-        Whether to actually create the folders specified. If False, only a
-        path will be generated but no folders will be created.
-    overwrite : bool
-        How to handle overwriting previously generated data.
-        If overwrite == False then no existing folders will be removed, however
-        if overwrite == True then any existing folders at the session level
-        or lower will be removed, including any contained data.
-    verbose : bool
-        If verbose is True, print status updates
-        as folders are created.
-
-    Returns
-    -------
-    path : str
-        The (relative) path to the folder that was created.
-
-    Examples
-    --------
-    >>> print(make_bids_folders('sub_01', session='my_session',
-                                kind='meg', bids_root='path/to/project',
-                                make_dir=False))  # noqa
-    path/to/project/sub-sub_01/ses-my_session/meg
-
-    """
-    _check_types((subject, kind, session, bids_root))
-    if session is not None:
-        _check_key_val('ses', session)
-
-    path = ['sub-%s' % subject]
-    if isinstance(session, str):
-        path.append('ses-%s' % session)
-    if isinstance(kind, str):
-        path.append(kind)
-    path = op.join(*path)
-    if isinstance(bids_root, str):
-        path = op.join(bids_root, path)
-
-    if make_dir is True:
-        _mkdir_p(path, overwrite=overwrite, verbose=verbose)
-    return path
 
 
 def make_dataset_description(path, name=None, data_license=None,
